refactor: home screen code organization
